# pkg/trace/api: never drop traces