--- a/pkg/trace/metrics/metrics.go
+++ b/pkg/trace/metrics/metrics.go
@@ -4,6 +4,7 @@ package metrics
 
 import (
 	"fmt"
+	"time"
 
 	"github.com/DataDog/datadog-agent/pkg/trace/config"
 	"github.com/DataDog/datadog-go/statsd"
@@ -14,6 +15,7 @@ type StatsClient interface {
 	Gauge(name string, value float64, tags []string, rate float64) error
 	Count(name string, value int64, tags []string, rate float64) error
 	Histogram(name string, value float64, tags []string, rate float64) error
+	Timing(name string, value time.Duration, tags []string, rate float64) error
 }
 
 // Client is a global Statsd client. When a client is configured via Configure,
@@ -42,6 +44,14 @@ func Histogram(name string, value float64, tags []string, rate float64) error {
 		return nil // no-op
 	}
 	return Client.Histogram(name, value, tags, rate)
+}
+
+// Timing calls Timing on the global Client, if set.
+func Timing(name string, value time.Duration, tags []string, rate float64) error {
+	if Client == nil {
+		return nil // no-op
+	}
+	return Client.Timing(name, value, tags, rate)
 }
 
 // Configure creates a statsd client for the given agent's configuration, using the specified global tags.