--- a/pkg/trace/watchdog/net_windows.go
+++ b/pkg/trace/watchdog/net_windows.go
@@ -1,6 +1,0 @@
-package watchdog
-
-// Net for windows returns basic network info without the number of connections.
-func (pi *CurrentInfo) Net() NetInfo {
-	return NetInfo{}
-}