… pkg/trace/test/testutil/statsd.go
… pkg/trace/test/testutil/trace.go
… pkg/trace/watchdog/info_nix_test.go
… pkg/trace/watchdog/info_test.go
… pkg/trace/watchdog/net.go
… releasenotes/notes/apm-never-drop-traces-bb6f2e36e8979070.yaml