This change modifies the HTTP receiver to never drop traces. Dropping
traces does not bring any real benefit. If a user wishes to process a
high amount of traces, he should be able to do so at the cost of
hardware resources. For now, we are safeguarded against excessive use of
memory and cpu by the watchdog system.

Additionally:
* Move the "watchdog" loop to the receiver
* Do not act on open network connection (this was a legacy protection no longer needed). It is superseded by `apm_config.connection_limit`
* Do not kill the agent when going above "max_memory". Instead, stop accepting requests and warn the user to increase the threshold.
* Switch the "watchdog check" interval to 20 seconds.