@gbbr 2019-04-02
@LotharSee I changed this as discussed to move watchdog to the HTTP receiver. Now it simply refuses requests when memory allocations go over `apm_config.max_memory`.

**Pros:**
* Traces are never dropped as long as the user is happy to increase resource allowance to accommodate.

**Cons:**
* We don't drop traces if memory overflows, we now drop the whole request (but at least we have a solution to not drop anything).