@gbbr 2019-04-03
@AlexJF thanks for looking. That's quite an interesting thought. How would that work though? One way that comes to mind, off the top of my head is to use a context with a deadline.

What you're saying is probably the most correct way to do it. It adds a bit more complexity but it might be significantly better. Although, I think I didn't quite describe the point of this change. The main point here is that if the user needs to allocate more memory to the trace-agent in order to process more traces, he gets notified of that by means of a 4xx HTTP response to his tracer and a warning in his logs. He would then immediately increase the value of "max_memory".

All in all, nobody wants to be in the situation where things are being dropped, so I think it's not worth going through all that trouble at this point.