@LotharSee 2019-04-23
@gbbr yes as long as we don't forget them I'm fine having these as part of another PR as these are not essential/blocking for the merge of this one.